Move packager into its own package
- This enables other libraries to import the packager's functionality

[#164880949]
