+++ packager/cnbpackager/packager.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package main
+package cnbpackager
 
 import (
 	"archive/tar"
